Refactoring: signOut
