+++ src/components/layout/header/MenuProfile.tsx
@@ -1,7 +1,7 @@
 import { useUser } from '@/hooks/useAPI';
 import profile from '@/components/layout/header/Profile.module.scss';
 import { createDefaultProfileImage } from '@/utils/createDefaultProfileImage';
-import { useSignOut } from '@/hooks/useSignOut';
+import { useSignOut } from '@/hooks/useAPI';
 import MyPageLink from '@/components/common/link/MyPageLink';
 
 interface MenuProfileProps {}
