+++ src/components/owner/Sidebar.tsx
@@ -1,15 +1,11 @@
 import NavChild from '@/components/common/link/NavChild';
 import owner from '@/components/owner/Owner.module.scss';
-import { useUser } from '@/hooks/useAPI';
 import Accordion from '@/components/common/accordion/Accordion';
 import MyPageLink from '@/components/common/link/MyPageLink';
+import { useParams } from 'react-router-dom';
 
 const OwnerSidebar = () => {
-  const { myData } = useUser();
-
-  if (myData === undefined) {
-    return null;
-  }
+  const { ownerId } = useParams<{ ownerId: string }>();
 
   return (
     <aside className={owner['sidebar']}>
@@ -17,18 +13,18 @@ const OwnerSidebar = () => {
       <MyPageLink>마이페이지(오너)</MyPageLink>
       <Accordion title="헬스장">
         <ul>
-          <NavChild to={`/owner/${myData!.id}/gym`} content="헬스장 현황" />
+          <NavChild to={`/owner/${ownerId}/gym`} content="헬스장 현황" />
         </ul>
       </Accordion>
       <Accordion title="스태프">
         <ul>
-          <NavChild to={`/owner/${myData!.id}/staff`} content="스태프 현황" />
+          <NavChild to={`/owner/${ownerId}/staff`} content="스태프 현황" />
         </ul>
       </Accordion>
       <Accordion title="멤버">
         <ul>
-          <NavChild to={`/owner/${myData!.id}/member`} content="멤버 현황" />
-          <NavChild to={`/owner/${myData!.id}/member/edit`} content="멤버 관리" />
+          <NavChild to={`/owner/${ownerId}/member`} content="멤버 현황" />
+          <NavChild to={`/owner/${ownerId}/member/edit`} content="멤버 관리" />
         </ul>
       </Accordion>
     </aside>
